quantech_test/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg[binary]==3.2.13
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-redis==5.4.0
django-filter==23.5
python-dateutil==2.8.2
pytz==2023.3
# Pillow==10.1.0
whitenoise==6.6.0
dj-database-url==2.1.0